refactor(restructure): dataset and controllers submodules
Move everything relevant into submodules, as per review feedback at https://github.com/cryoem-uoft/cryosparc-tools/pull/114#issuecomment-2573571747
