+++ cryosparc/dataset/core.pyi
